avances de clase

+++ Point.h
@@ -5,9 +5,9 @@ using namespace std;
 struct Point{
     int x;
     int y;
-    int value;
+    double value;
 
     Point() {}
 
-    Point(int x, int y, int value) : x(x), y(y), value(value){}
+    Point(int x, int y, double value) : x(x), y(y), value(value){}
 };
